Generates an Installation Qualification Report

The run_iq function executes an installation qualification (IQ) on the currently installed sassy packages, and generates a report on the results. The IQ ensures that all sassy packages have been installed. The results of the IQ will be placed in the supplied location.

run_iq(location)

Arguments

location

The path to the desired output directory. The IQ reports and any associated files will be placed in this directory.

Value

The path to the output directory. The directory will contain a PDF report showing a summary of the results of the IQ. After the function is run, review this report to ensure that all tests passed.

Examples

# Create a temp directory
tmp <- tempdir()

# Run the Installation Qualification
run_iq(tmp)