sassy: Making R Easier for SAS® Programmers

The sassy package is a meta-package that installs a set of functions that aim to make R easier for programmers whose primary experience is with SAS® software. These functions provide the ability to create data libraries, format catalogs, data dictionaries, a traceable log, and includes reporting capabilities reminiscent of those found in SAS®. These packages were written independently, and the authors have no association with, approval of, or endorsement by SAS® Institute.

Packages Included

The packages included in the sassy meta-package are as follows:

libr: Define a libname, view data dictionaries, and simulate a data step.

fmtr: Create a format catalog and apply formats to a data frame or vector.

logr: Generate a traceable log.

reporter: Write reports and output in text, RTF, and PDF.

Examples

#########################################################
#####         Example: Simple Data Listing          #####
#########################################################
library(sassy)
library(magrittr)

options("logr.notes" = FALSE)

# Get path to temp directory
tmp <- tempdir()

# Get path to sample data
pkg <- system.file("extdata", package = "sassy")

# Open log
lgpth <- log_open(file.path(tmp, "example1.log"))

sep("Get Data")

#> [1] "========================================================================="
#> [1] "Get Data"
#> [1] "========================================================================="

# Define data library
libname(sdtm, pkg, "csv") %>% put()

#> $AE

#> 
#> -- Column specification --------------------------------------------------------
#> cols(
#>   .default = col_character(),
#>   AESEQ = col_double(),
#>   AELLT = col_logical(),
#>   AELLTCD = col_logical(),
#>   AEPTCD = col_double(),
#>   AEHLTCD = col_double(),
#>   AEHLGTCD = col_double(),
#>   AEBDSYCD = col_double(),
#>   AESOCCD = col_double(),
#>   AESTDY = col_double(),
#>   AEENDY = col_double()
#> )
#> i Use `spec()` for the full column specifications.

#> $DA

#> 
#> -- Column specification --------------------------------------------------------
#> cols(
#>   STUDYID = col_character(),
#>   DOMAIN = col_character(),
#>   USUBJID = col_character(),
#>   DASEQ = col_double(),
#>   DAREFID = col_character(),
#>   DATESTCD = col_character(),
#>   DATEST = col_character(),
#>   DAORRES = col_double(),
#>   DAORRESU = col_character(),
#>   DASTRESC = col_double(),
#>   DASTRESN = col_double(),
#>   DASTRESU = col_character(),
#>   DASTAT = col_character(),
#>   DADRVFL = col_character(),
#>   VISITNUM = col_double(),
#>   VISIT = col_character(),
#>   DADTC = col_date(format = ""),
#>   DADY = col_double()
#> )

#> $DM

#> 
#> -- Column specification --------------------------------------------------------
#> cols(
#>   .default = col_character(),
#>   RFSTDTC = col_date(format = ""),
#>   RFENDTC = col_date(format = ""),
#>   RFXSTDTC = col_logical(),
#>   RFXENDTC = col_logical(),
#>   RFICDTC = col_date(format = ""),
#>   RFPENDTC = col_date(format = ""),
#>   DTHDTC = col_logical(),
#>   DTHFL = col_logical(),
#>   BRTHDTC = col_date(format = ""),
#>   AGE = col_double(),
#>   ACTARMCD = col_logical(),
#>   ACTARM = col_logical(),
#>   COUNTRY = col_logical()
#> )
#> i Use `spec()` for the full column specifications.

#> $DS

#> 
#> -- Column specification --------------------------------------------------------
#> cols(
#>   STUDYID = col_character(),
#>   DOMAIN = col_character(),
#>   USUBJID = col_character(),
#>   DSSEQ = col_double(),
#>   DSTERM = col_character(),
#>   DSDECOD = col_character(),
#>   DSCAT = col_character(),
#>   DSSTDTC = col_character(),
#>   DSSTDY = col_double()
#> )

#> $EX

#> 
#> -- Column specification --------------------------------------------------------
#> cols(
#>   STUDYID = col_character(),
#>   DOMAIN = col_character(),
#>   USUBJID = col_character(),
#>   EXSEQ = col_double(),
#>   EXTRT = col_character(),
#>   EXDOSE = col_logical(),
#>   EXDOSTXT = col_character(),
#>   EXDOSU = col_character(),
#>   EXDOSFRM = col_character(),
#>   EXSTDTC = col_date(format = ""),
#>   EXENDTC = col_character()
#> )

#> $IE

#> 
#> -- Column specification --------------------------------------------------------
#> cols(
#>   STUDYID = col_character(),
#>   DOMAIN = col_character(),
#>   USUBJID = col_character(),
#>   IESEQ = col_double(),
#>   IETESTCD = col_character(),
#>   IETEST = col_character(),
#>   IECAT = col_character(),
#>   IEORRES = col_character(),
#>   IESTRESC = col_character(),
#>   VISITNUM = col_double(),
#>   VISIT = col_character(),
#>   VISITDY = col_double(),
#>   IEDTC = col_date(format = ""),
#>   IEDY = col_double()
#> )

#> $SV

#> 
#> -- Column specification --------------------------------------------------------
#> cols(
#>   STUDYID = col_character(),
#>   DOMAIN = col_character(),
#>   USUBJID = col_character(),
#>   VISITNUM = col_double(),
#>   VISIT = col_character(),
#>   SVSTDTC = col_date(format = ""),
#>   SVENDTC = col_date(format = ""),
#>   SVSTDY = col_double(),
#>   SVENDY = col_double(),
#>   SVUPDES = col_character()
#> )

#> $VS

#> 
#> -- Column specification --------------------------------------------------------
#> cols(
#>   STUDYID = col_character(),
#>   DOMAIN = col_character(),
#>   USUBJID = col_character(),
#>   VSSEQ = col_double(),
#>   VSTESTCD = col_character(),
#>   VSTEST = col_character(),
#>   VSPOS = col_character(),
#>   VSORRES = col_double(),
#>   VSORRESU = col_character(),
#>   VSSTRESC = col_double(),
#>   VSSTRESN = col_double(),
#>   VSSTRESU = col_character(),
#>   VSBLFL = col_character(),
#>   VISITNUM = col_double(),
#>   VISIT = col_character(),
#>   VSDTC = col_date(format = ""),
#>   VSDY = col_double()
#> )

#> # library 'sdtm': 8 items
#> - attributes: csv not loaded
#> - path: C:/Users/User/AppData/Local/Temp/RtmpmUMDym/temp_libpath30206b32a06/sassy/extdata
#> - items:
#>   Name Extension Rows Cols     Size        LastModified
#> 1   AE       csv  150   27  88.2 Kb 2020-12-21 09:44:51
#> 2   DA       csv 3587   18   528 Kb 2020-12-21 09:44:51
#> 3   DM       csv   87   24  45.3 Kb 2020-12-21 09:44:51
#> 4   DS       csv  174    9  33.8 Kb 2020-12-21 09:44:51
#> 5   EX       csv   84   11  26.1 Kb 2020-12-21 09:44:51
#> 6   IE       csv    2   14  13.1 Kb 2020-12-21 09:44:51
#> 7   SV       csv  685   10  70.1 Kb 2020-12-21 09:44:51
#> 8   VS       csv 3358   17 467.2 Kb 2020-12-21 09:44:51

# Load library into workspace
lib_load(sdtm)

#> # library 'sdtm': 8 items
#> - attributes: csv loaded
#> - path: C:/Users/User/AppData/Local/Temp/RtmpmUMDym/temp_libpath30206b32a06/sassy/extdata
#> - items:
#>   Name Extension Rows Cols     Size        LastModified
#> 1   AE       csv  150   27  88.2 Kb 2020-12-21 09:44:51
#> 2   DA       csv 3587   18   528 Kb 2020-12-21 09:44:51
#> 3   DM       csv   87   24  45.3 Kb 2020-12-21 09:44:51
#> 4   DS       csv  174    9  33.8 Kb 2020-12-21 09:44:51
#> 5   EX       csv   84   11  26.1 Kb 2020-12-21 09:44:51
#> 6   IE       csv    2   14  13.1 Kb 2020-12-21 09:44:51
#> 7   SV       csv  685   10  70.1 Kb 2020-12-21 09:44:51
#> 8   VS       csv 3358   17 467.2 Kb 2020-12-21 09:44:51

sep("Write Report")

#> [1] "========================================================================="
#> [1] "Write Report"
#> [1] "========================================================================="

# Define table object
tbl <- create_table(sdtm.DM) %>%
  define(USUBJID, id_var = TRUE)

# Construct report path
pth <- file.path(tmp, "output/l_dm.rtf") %>% put()

#> [1] "C:\\Users\\User\\AppData\\Local\\Temp\\RtmpIN0Jrr/output/l_dm.rtf"

# Define report object
rpt <- create_report(pth, output_type = "RTF") %>%
  page_header("Sponsor: Company", "Study: ABC") %>%
  titles("Listing 1.0", "SDTM Demographics") %>%
  add_content(tbl, align = "left") %>%
  page_footer(Sys.time(), "CONFIDENTIAL", "Page [pg] of [tpg]")

# Write report to file system
write_report(rpt) %>% put()

#> # A report specification: 9 pages
#> - file_path: 'C:\Users\User\AppData\Local\Temp\RtmpIN0Jrr/output/l_dm.rtf'
#> - output_type: RTF
#> - units: inches
#> - orientation: landscape
#> - line size/count: 108/48
#> - page_header: left=Sponsor: Company right=Study: ABC
#> - title 1: 'Listing 1.0'
#> - title 2: 'SDTM Demographics'
#> - page_footer: left=2020-12-21 09:45:00 center=CONFIDENTIAL right=Page [pg] of [tpg]
#> - content: 
#> # A table specification:
#> - data: tibble 'sdtm.DM' 87 rows 24 cols
#> - show_cols: all
#> - use_attributes: all
#> - define: USUBJID id_var='TRUE' 

# Unload data
lib_unload(sdtm)

#> # library 'sdtm': 8 items
#> - attributes: csv not loaded
#> - path: C:/Users/User/AppData/Local/Temp/RtmpmUMDym/temp_libpath30206b32a06/sassy/extdata
#> - items:
#>   Name Extension Rows Cols     Size        LastModified
#> 1   AE       csv  150   27  88.2 Kb 2020-12-21 09:44:51
#> 2   DA       csv 3587   18   528 Kb 2020-12-21 09:44:51
#> 3   DM       csv   87   24  45.3 Kb 2020-12-21 09:44:51
#> 4   DS       csv  174    9  33.8 Kb 2020-12-21 09:44:51
#> 5   EX       csv   84   11  26.1 Kb 2020-12-21 09:44:51
#> 6   IE       csv    2   14  13.1 Kb 2020-12-21 09:44:51
#> 7   SV       csv  685   10  70.1 Kb 2020-12-21 09:44:51
#> 8   VS       csv 3358   17 467.2 Kb 2020-12-21 09:44:51

# Close log
log_close()

# Print log to console
writeLines(readLines(lgpth, encoding = "UTF-8"))

#> ========================================================================= 
#> Log Path: C:/Users/User/AppData/Local/Temp/RtmpIN0Jrr/log/example1.log 
#> Working Directory: C:/packages/sassy/docs/reference 
#> User Name: User 
#> R Version: 4.0.3 (2020-10-10) 
#> Machine: ZB15-5CD00276ZY x86-64 
#> Operating System: Windows 10 x64 build 18363 
#> Log Start Time: 2020-12-21 09:44:59 
#> ========================================================================= 
#> 
#> ========================================================================= 
#> Get Data 
#> ========================================================================= 
#> 
#> # library 'sdtm': 8 items
#> - attributes: csv not loaded
#> - path: C:/Users/User/AppData/Local/Temp/RtmpmUMDym/temp_libpath30206b32a06/sassy/extdata
#> - items:
#>   Name Extension Rows Cols     Size        LastModified
#> 1   AE       csv  150   27  88.2 Kb 2020-12-21 09:44:51
#> 2   DA       csv 3587   18   528 Kb 2020-12-21 09:44:51
#> 3   DM       csv   87   24  45.3 Kb 2020-12-21 09:44:51
#> 4   DS       csv  174    9  33.8 Kb 2020-12-21 09:44:51
#> 5   EX       csv   84   11  26.1 Kb 2020-12-21 09:44:51
#> 6   IE       csv    2   14  13.1 Kb 2020-12-21 09:44:51
#> 7   SV       csv  685   10  70.1 Kb 2020-12-21 09:44:51
#> 8   VS       csv 3358   17 467.2 Kb 2020-12-21 09:44:51
#> 
#> ========================================================================= 
#> Write Report 
#> ========================================================================= 
#> 
#> C:\Users\User\AppData\Local\Temp\RtmpIN0Jrr/output/l_dm.rtf 
#> 
#> # A report specification: 9 pages
#> - file_path: 'C:\Users\User\AppData\Local\Temp\RtmpIN0Jrr/output/l_dm.rtf'
#> - output_type: RTF
#> - units: inches
#> - orientation: landscape
#> - line size/count: 108/48
#> - page_header: left=Sponsor: Company right=Study: ABC
#> - title 1: 'Listing 1.0'
#> - title 2: 'SDTM Demographics'
#> - page_footer: left=2020-12-21 09:45:00 center=CONFIDENTIAL right=Page [pg] of [tpg]
#> - content: 
#> # A table specification:
#> - data: tibble 'sdtm.DM' 87 rows 24 cols
#> - show_cols: all
#> - use_attributes: all
#> - define: USUBJID id_var='TRUE' 
#> 
#> ========================================================================= 
#> Log End Time: 2020-12-21 09:45:02 
#> Log Elapsed Time: 0 00:00:02 
#> ========================================================================= 

# View report
# file.show(pth)