sassy: Making R Easier for SAS® Programmers

The sassy package is a meta-package that installs a set of functions that aim to make R easier for programmers whose primary experience is with SAS® software. These functions provide the ability to create data libraries, format catalogs, data dictionaries, a traceable log, and includes reporting capabilities reminiscent of those found in SAS®. These packages were written independently, and the authors have no association with, approval of, or endorsement by SAS® Institute.

Packages Included

The packages included in the sassy meta-package are as follows:

libr: Define a libname, view data dictionaries, and simulate a data step.

fmtr: Create a format catalog and apply formats to a data frame or vector.

logr: Generate a traceable log.

reporter: Write reports and output in text, RTF, and PDF.

Examples

#########################################################
#####         Example: Simple Data Listing          #####
#########################################################
library(sassy)

options("logr.notes" = FALSE)

# Get path to temp directory
tmp <- tempdir()

# Get path to sample data
pkg <- system.file("extdata", package = "sassy")

# Open log
lgpth <- log_open(file.path(tmp, "example1.log"))

sep("Get Data")

#> [1] "========================================================================="
#> [1] "Get Data"
#> [1] "========================================================================="

# Define data library
libname(sdtm, pkg, "csv") %>% put()

#> Error in libname(sdtm, pkg, "csv") %>% put(): could not find function "%>%"

# # library 'sdtm': 10 items
# - attributes: csv not loaded
# - path: C:/Users/User/Documents/R/win-library/4.0/sassy/extdata
# - items:
#   Name Extension Rows Cols     Size        LastModified
# 1      AE       csv  150   27  88.1 Kb 2020-12-17 19:05:00
# 2      DA       csv 3587   18 527.8 Kb 2020-12-17 19:05:00
# 3      DM       csv   87   24  45.2 Kb 2020-12-17 19:05:00
# 4      DS       csv  174    9  33.7 Kb 2020-12-17 19:05:00
# 5      EX       csv   84   11    26 Kb 2020-12-17 19:05:00
# 6      IE       csv    2   14    13 Kb 2020-12-17 19:05:00
# 7      PE       csv 1854   17 277.9 Kb 2020-12-17 19:05:00
# 8  SUPPEX       csv  639   10  63.7 Kb 2020-12-17 19:05:00
# 9      SV       csv  685   10  69.9 Kb 2020-12-17 19:05:00
# 10     VS       csv 3358   17   467 Kb 2020-12-17 19:05:00

# Load library into workspace
lib_load(sdtm)

#> Error in lib_load(sdtm): object 'sdtm' not found

sep("Write Report")

#> [1] "========================================================================="
#> [1] "Write Report"
#> [1] "========================================================================="

# Define table object
tbl <- create_table(sdtm.DM) %>%
  define(USUBJID, id_var = TRUE)

#> Error in create_table(sdtm.DM) %>% define(USUBJID, id_var = TRUE): could not find function "%>%"

# Construct report path
pth <- file.path(tmp, "output/l_dm.rtf") %>% put()

#> Error in file.path(tmp, "output/l_dm.rtf") %>% put(): could not find function "%>%"

# Define report object
rpt <- create_report(pth, output_type = "RTF") %>%
  page_header("Sponsor: Company", "Study: ABC") %>%
  titles("Listing 1.0", "SDTM Demographics") %>%
  add_content(tbl, align = "left") %>%
  page_footer(Sys.time(), "CONFIDENTIAL", "Page [pg] of [tpg]")

#> Error in create_report(pth, output_type = "RTF") %>% page_header("Sponsor: Company",     "Study: ABC") %>% titles("Listing 1.0", "SDTM Demographics") %>%     add_content(tbl, align = "left") %>% page_footer(Sys.time(),     "CONFIDENTIAL", "Page [pg] of [tpg]"): could not find function "%>%"

# Write report to file system
write_report(rpt) %>% put()

#> Error in write_report(rpt) %>% put(): could not find function "%>%"

# # A report specification: 9 pages
# - file_path: 'C:\Users\User\AppData\Local\Temp\RtmpsRnVcf/output/l_dm.rtf'
# - output_type: RTF
# - units: inches
# - orientation: landscape
# - line size/count: 108/48
# - page_header: left=Sponsor: Company right=Study: ABC
# - title 1: 'Listing 1.0'
# - title 2: 'SDTM Demographics'
# - page_footer: left=2020-12-13 17:27:41 center=CONFIDENTIAL right=Page [pg] of [tpg]
# - content:
# # A table specification:
# - data: tibble 'sdtm.DM' 87 rows 24 cols
# - show_cols: all
# - use_attributes: all
# - define: USUBJID id_var='TRUE'

# Unload data
lib_unload(sdtm)

#> Error in lib_unload(sdtm): object 'sdtm' not found

# Close log
log_close()

# Print log to console
writeLines(readLines(lgpth, encoding = "UTF-8"))

#> ========================================================================= 
#> Log Path: C:/Users/User/AppData/Local/Temp/Rtmp2vbEBy/log/example1.log 
#> Working Directory: C:/packages/sassy/docs/reference 
#> User Name: User 
#> R Version: 4.0.3 (2020-10-10) 
#> Machine: ZB15-5CD00276ZY x86-64 
#> Operating System: Windows 10 x64 build 18363 
#> Log Start Time: 2020-12-18 22:14:57 
#> ========================================================================= 
#> 
#> ========================================================================= 
#> Get Data 
#> ========================================================================= 
#> 
#> ========================================================================= 
#> Write Report 
#> ========================================================================= 
#> 
#> ========================================================================= 
#> Log End Time: 2020-12-18 22:14:57 
#> Log Elapsed Time: 0 00:00:00 
#> ========================================================================= 

# View report
# file.show(pth)